Adjusted timer and refactored code
The timer must not be stopped; otherwise, the weather widget will only be updated once and only after installing it or modifying its settings (this is something I did not realise until I tested and debugged the watchface on the band itself, as I was testing this specific feature on ZeppPlayer). Now it will run approximately twice per each default screen timeout, to ensure the assets get updated if daytime/nighttime is detected.

Also, I moved  to  as it did not fit in  and refactored it creating global constants for the sensors and global variable to keep the previous daytime/nighttime status. I noticed multiple sensors were created when calling , which was not optimal and could cause potential battery drains.